added new control ListMenu - added custom draw entry func improved GetUIAnchor function added strings functionality improved asset data encryption - now includes filenames and size cleanups

+++ src/internal/utils/EncryptionUtil.h
@@ -5,7 +5,7 @@
 // Simple xor encryption with a 64-bit key - should generally be enough
 // https://reverseengineering.stackexchange.com/questions/2062/what-is-the-most-efficient-way-to-detect-and-to-break-xor-encryption
 
-inline void SymmetricEncrypt(char* data, const int size, const unsigned long long key)
+inline void SymmetricEncrypt(char* data, const int size, const uint64_t key)
 {
     for (int i = 0; i < size; ++i)
     {
